+++ src/app/user/[slug]/UserContent.jsx
@@ -8,9 +8,9 @@ import PropTypes from "prop-types";
 import "@/app/globals.css";
 import GenrePieChart from "@/shared/GenrePieChart";
 import Boombox from "@/app/user/[slug]/Boombox";
-import Cassette from "@/app/user/[slug]/Cassette";
+import Cassette from "@/shared/Cassette";
 import PaperTitle from "@/app/user/[slug]/PaperTitle";
-import CDStack from "@/app/user/[slug]/CDStack";
+import CDStack from "@/shared/CDStack";
 import getPersonality from "@/shared/GetPersonality";
 import PhotoMarquee from "@/app/user/[slug]/PhotoMarquee";
 import PhotoGallery from "@/app/user/[slug]/PhotoGallery";
